third_party/blink/web_tests/platform/generic/svg/dom/SVGViewSpec-expected.txt

CONSOLE ERROR: Uncaught SyntaxError: Unexpected end of input
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
